+++ src/audio_metadata/formats/vorbis.py
@@ -1,7 +1,6 @@
 __all__ = [
 	'VorbisComment',
 	'VorbisComments',
-	'VorbisPicture',
 ]
 
 import struct
@@ -10,12 +9,8 @@ from collections import defaultdict
 from attr import attrib, attrs
 from tbm_utils import AttrMapping
 
-from .tables import ID3PictureType
 from ..exceptions import InvalidComment
-from ..models import (
-	Picture,
-	Tags,
-)
+from ..models import Tags
 from ..utils import datareader
 
 
@@ -54,30 +49,3 @@ class VorbisComments(Tags):
 			fields[comment.name].append(comment.value)
 
 		return cls(**fields, _vendor=vendor)
-
-
-class VorbisPicture(Picture):
-	@datareader
-	@classmethod
-	def load(cls, data):
-		type_, mime_length = struct.unpack('>2I', data.read(8))
-		mime_type = data.read(mime_length).decode('utf-8', 'replace')
-
-		desc_length = struct.unpack('>I', data.read(4))[0]
-		description = data.read(desc_length).decode('utf-8', 'replace')
-
-		width, height, depth, colors = struct.unpack('>4I', data.read(16))
-
-		data_length = struct.unpack('>I', data.read(4))[0]
-		data = data.read(data_length)
-
-		return cls(
-			type=ID3PictureType(type_),
-			mime_type=mime_type,
-			description=description,
-			width=width,
-			height=height,
-			depth=depth,
-			colors=colors,
-			data=data,
-		)