+++ tests/test_formats_flac.py
@@ -8,10 +8,12 @@ from audio_metadata import (
 	FLACCueSheetTrack,
 	FLACMetadataBlock,
 	FLACPadding,
+	FLACPicture,
 	FLACSeekPoint,
 	FLACSeekTable,
+	ID3PictureType,
 	InvalidBlock,
-	InvalidHeader
+	InvalidHeader,
 )
 
 
@@ -320,6 +322,33 @@ def test_FLACPadding():
 
 	assert padding_init == padding_load
 	assert repr(padding_init) == repr(padding_load) == '<FLACPadding (10 bytes)>'
+
+
+def test_FLACPicture():
+	image_data = (
+		b'\x00\x00\x00\x03\x00\x00\x00\timage/png\x00\x00\x00\x00'
+		b'\x00\x00\x00\x10\x00\x00\x00\x10\x00\x00\x00 \x00\x00'
+		b'\x00\x00\x00\x00\x00`\x89PNG\r\n\x1a\n\x00\x00\x00\r'
+		b'IHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x06\x00\x00'
+		b'\x00\x1f\xf3\xffa\x00\x00\x00\tpHYs\x00\x00\x0b\x12\x00'
+		b'\x00\x0b\x12\x01\xd2\xdd~\xfc\x00\x00\x00\x12IDAT8\xcbc`'
+		b'\x18\x05\xa3`\x14\x8c\x02\x08\x00\x00\x04\x10\x00\x01\x85'
+		b'?\xaar\x00\x00\x00\x00IEND\xaeB`\x82'
+	)
+
+	vorbis_picture_init = FLACPicture(
+		type=ID3PictureType.COVER_FRONT,
+		mime_type='image/png',
+		description='',
+		width=16,
+		height=16,
+		depth=32,
+		colors=0,
+		data=image_data[41:],
+	)
+	vorbis_picture_load = FLACPicture.load(image_data)
+
+	assert vorbis_picture_init == vorbis_picture_load
 
 
 def test_FLACSeektable():